fix: move `formatDate()` to `utils/time`
