+++ site/src/pages/DeploymentSettingsPage/OverviewPage/UserEngagementChart.tsx
@@ -16,7 +16,7 @@ import { ChevronRightIcon } from "lucide-react";
 import type { FC } from "react";
 import { Link as RouterLink } from "react-router";
 import { Area, AreaChart, CartesianGrid, XAxis, YAxis } from "recharts";
-import { formatDate } from "utils/date";
+import { formatDate } from "utils/time";
 
 const chartConfig = {
 	users: {
