+++ site/src/pages/WorkspacePage/WorkspaceTopbar.tsx
@@ -30,7 +30,7 @@ import { WorkspaceStatusIndicator } from "modules/workspaces/WorkspaceStatusIndi
 import type { FC } from "react";
 import { useQuery } from "react-query";
 import { Link as RouterLink } from "react-router";
-import { formatDate } from "utils/date";
+import { formatDate } from "utils/time";
 import { displayDormantDeletion } from "utils/dormant";
 import type { WorkspacePermissions } from "../../modules/workspaces/permissions";
 import { WorkspaceActions } from "./WorkspaceActions/WorkspaceActions";
